Get data in Details parent component
